GH-1780 [Bug] 修复bash脚本超时异常 #1784

+++ spark-yun-backend/spark-yun-modules/src/main/java/com/isxcode/star/modules/datasource/service/biz/DatasourceBizService.java
@@ -23,6 +23,7 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.sql.Connection;
@@ -272,6 +273,17 @@ public class DatasourceBizService {
             }
         }
 
+        // 判断名称重复
+        if (databaseDriverRepository.existsDatabaseDriverEntitiesByName(name)) {
+            throw new IsxAppException("驱动名称重复");
+        }
+
+        // 判断文件是否已经存在
+        Path path = Paths.get(driverDirPath + File.separator + driverFile.getOriginalFilename());
+        if (path.toFile().exists()) {
+            throw new IsxAppException("该驱动文件已存在");
+        }
+
         // 保存驱动文件
         try (InputStream inputStream = driverFile.getInputStream()) {
             Files.copy(inputStream, Paths.get(driverDirPath).resolve(driverFile.getOriginalFilename()),
@@ -390,4 +402,12 @@ public class DatasourceBizService {
         return datasourceMapper.databaseDriverEntityToGetDefaultDatabaseDriverRes(systemDriver.get());
     }
 
+    public void updateDatabaseDriverRemark(UpdateDatabaseDriverRemarkReq updateDatabaseDriverRemarkReq) {
+
+        DatabaseDriverEntity datasourceDriver =
+            datasourceService.getDatasourceDriver(updateDatabaseDriverRemarkReq.getId());
+
+        datasourceDriver.setRemark(updateDatabaseDriverRemarkReq.getRemark());
+        databaseDriverRepository.save(datasourceDriver);
+    }
 }
